Image fixes + css fixes + Wording & punctuation Fixed a few hyperlinks for external images that were no longer available.
Fixed the styling for the rows of images on the items & monsters pages
so that there are always three images positioned on each row, and no gaps.

Other styling fixes such as adding a vertical divider on the "about us" page
for better readability, as well as a linebreak following the horizontal divider.
Also added padding on the left-hand side of the unordered list at the bottom of
the "tips and tricks" page for better positioning and readability.

Improved the internal hyperlinks on the "tips and tricks" page to point to the
exact item or monster being referenced rather than just the full page itself.
Also added a few extra ones.

Made extensive grammar, spelling, punctuation, and wording fixes across all pages.